refactor: accept only .smacc files and relocate file menu items
